+++ flask_rest_jsonapi_next/decorators.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+# isort: skip_file
+# fmt: off
 
 """Decorators to check headers and method requirements for each Api calls"""
 
@@ -7,9 +9,9 @@ from functools import wraps
 
 from flask import request, make_response, jsonify, current_app
 
-from flask_rest_jsonapi_next.errors import jsonapi_errors
-from flask_rest_jsonapi_next.exceptions import JsonApiException
-from flask_rest_jsonapi_next.utils import JSONEncoder
+from .errors import jsonapi_errors
+from .exceptions import JsonApiException
+from .utils import JSONEncoder
 
 
 def check_headers(func):
@@ -21,14 +23,19 @@ def check_headers(func):
     @wraps(func)
     def wrapper(*args, **kwargs):
         if request.method in ('POST', 'PATCH'):
-            if 'Content-Type' not in request.headers or\
-                    'application/vnd.api+json' not in request.headers['Content-Type'] or\
-                    request.headers['Content-Type'] != 'application/vnd.api+json':
-                error = json.dumps(jsonapi_errors([{'source': '',
-                                                    'detail': "Content-Type header must be application/vnd.api+json",
-                                                    'title': 'Invalid request header',
-                                                    'status': '415'}]), cls=JSONEncoder)
-                return make_response(error, 415, {'Content-Type': 'application/vnd.api+json'})
+            if (
+                'Content-Type' not in request.headers
+                or (
+                    request.mimetype != 'application/vnd.api+json'
+                    and request.mimetype != 'application/json'
+                )
+            ):
+                raise JsonApiException(
+                    detail="Content-Type header must be application/vnd.api+json or application/json",
+                    title="Invalid request header",
+                    status=415,
+                )
+
         if 'Accept' in request.headers:
             flag = False
             for accept in request.headers['Accept'].split(','):
@@ -38,12 +45,12 @@ def check_headers(func):
                 if 'application/vnd.api+json' in accept and accept.strip() != 'application/vnd.api+json':
                     flag = True
             if flag is True:
-                error = json.dumps(jsonapi_errors([{'source': '',
-                                                    'detail': ('Accept header must be application/vnd.api+json without'
-                                                               'media type parameters'),
-                                                    'title': 'Invalid request header',
-                                                    'status': '406'}]), cls=JSONEncoder)
-                return make_response(error, 406, {'Content-Type': 'application/vnd.api+json'})
+                raise JsonApiException(
+                    detail="Accept header must be application/vnd.api+json without media type parameters",
+                    title="Invalid request header",
+                    status=406,
+                )
+
         return func(*args, **kwargs)
     return wrapper
 
@@ -68,35 +75,4 @@ def check_method_requirements(func):
         return func(*args, **kwargs)
     return wrapper
 
-
-def jsonapi_exception_formatter(func):
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        headers = {'Content-Type': 'application/vnd.api+json'}
-        try:
-            return func(*args, **kwargs)
-        except JsonApiException as e:
-            return make_response(jsonify(jsonapi_errors([e.to_dict()])),
-                                 e.status,
-                                 headers)
-        except Exception as e:
-            if current_app.config['DEBUG'] is True or current_app.config.get('PROPAGATE_EXCEPTIONS') is True:
-                raise
-
-            if 'sentry' in current_app.extensions:
-                current_app.extensions['sentry'].captureException()
-
-            exc = JsonApiException(getattr(e,
-                                           'detail',
-                                           current_app.config.get('GLOBAL_ERROR_MESSAGE') or str(e)),
-                                   source=getattr(e, 'source', ''),
-                                   title=getattr(e, 'title', None),
-                                   status=getattr(e, 'status', None),
-                                   code=getattr(e, 'code', None),
-                                   id_=getattr(e, 'id', None),
-                                   links=getattr(e, 'links', None),
-                                   meta=getattr(e, 'meta', None))
-            return make_response(jsonify(jsonapi_errors([exc.to_dict()])),
-                                 exc.status,
-                                 headers)
-    return wrapper
+# fmt: on