+++ flask_rest_jsonapi_next/utils.py
@@ -1,17 +1,24 @@
 # -*- coding: utf-8 -*-
+# isort: skip_file
+# fmt: off
 
-import json
-from uuid import UUID
-from datetime import datetime
-from decimal import Decimal
+from flask import current_app
 
+from .json_utilities import json, JSONEncoder, _HAS_RAPIDJSON
 
-class JSONEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, datetime):
-            return obj.isoformat()
-        elif isinstance(obj, UUID):
-            return str(obj)
-        elif isinstance(obj, Decimal):
-            return str(obj)            
-        return json.JSONEncoder.default(self, obj)
+def json_dumps(obj):
+    sort_keys = current_app.config.get("JSON_SORT_KEYS", False)
+
+    if _HAS_RAPIDJSON:
+        return json.dumps(
+            obj,
+            sort_keys=sort_keys,
+            number_mode=json.NM_DECIMAL,
+            datetime_mode=json.DM_ISO8601,
+            uuid_mode=json.UM_CANONICAL
+        )
+
+    else:
+        return json.dumps(obj, sort_keys=sort_keys, cls=JSONEncoder)
+
+# fmt: on